Move YAML link files to data
